restructure Listing Form WIP
