+++ src/app/(create-listings)/create-listing/products/page.jsx
@@ -1,6 +1,6 @@
 "use client";
 import { ProtectedRoute } from "@/components/ProtectedRoute";
-import ListingForm from "@/components/ListingForm";
+import ListingForm from "@/components/ListingForm2";
 import { useAuthState } from "react-firebase-hooks/auth";
 import { auth } from "@/firebase";
 import categories from "./categories";
